+++ MeongMungI/Domain/Interfaces/CommunityRepositoryProtocol.swift
@@ -10,11 +10,11 @@ import Foundation
 import RxSwift
 
 // MARK: - 게시글 프로토콜(인터페이스)
-protocol PostRepositoryProtocol {
+protocol CommunityRepositoryProtocol {
     // MARK: - 게시글 생성
     func createPost()
     // MARK: - 게시글 조회
-    func fetchList() -> Single<[PostResponseDTO]>
+    func fetchList() -> Single<[PostDTO]>
     // MARK: - 게시글 수정
     func updatePost()
     // MARK: - 게시글 삭제
