+++ MeongMungI/Presentation/CommunityScene/ViewModel/PostListViewModel.swift
@@ -1,0 +1,30 @@
+//
+//  CommunityViewModel.swift
+//  MeongMungI
+//
+//  Created by 7aeHoon on 7/5/25.
+//
+
+import Foundation
+
+import RxSwift
+
+public final class PostListViewModel {
+    private let disposeBag = DisposeBag()
+    
+    public struct Input {
+        
+    }
+    
+    public struct Output {
+        
+    }
+    
+    
+    public func transform(input: Input) -> Output {
+        let input = Input()
+        
+        return Output()
+    }
+
+}
